potential fix for green poop display issue
I think I finally got to the bottom of the issue with the resource display: The text boxes used for tracking these values were just barely big enough, and I think "GREEN" takes up slightly more space than any of the other resource names. So  when your green count went into the double digits, it went out of the text box and caused the glitch, I think. So I increased the size of that text box in the prefab.