re-worked hit detection on player melee attack
in practice, it was proving rather difficult to hit enemies with the melee attack. Therefore, I changed how it determines target. Now instead of drawing a circle around a point, the code creates a rectangular area using two points as the corners. This allowed greater control of the area the attack can hit, and I was able to make a shape that feels much better to use.